comic_agent/requirements.txt
torch>=2.1.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
safetensors>=0.4.1
openai>=1.0.0
python-dotenv>=1.0.0
colorama>=0.4.6
Pillow>=10.0.0
gradio>=4.0.0
